:recycle: Move dataset files api into datasets namespace.
Why? because datasetId is a required prop.
